# expression: implement vectorized evaluation for `builtinUnaryMinusDecimalSig`