PCP #12102 

### What problem does this PR solve? <!--add issue link with summary if exists-->
Implement vectorized evaluation for `builtinUnaryMinusDecimalSig` at #12102 

### What is changed and how it works?
It gets faster about 3.8x.

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test
```
GO111MODULE=on go test -check.f TestVectorizedBuiltinOpFunc
PASS: builtin_op_vec_test.go:136: testEvaluatorSuite.TestVectorizedBuiltinOpFunc	0.043s
OK: 1 passed
PASS
ok  	github.com/pingcap/tidb/expression	0.066s
```

 - Benchmark
```
go test -v -benchmem -bench=BenchmarkVectorizedBuiltinOpFunc -run=BenchmarkVectorizedBuiltinOpFunc -args builtinUnaryMinusDecimalSig
goos: darwin
goarch: amd64
pkg: github.com/pingcap/tidb/expression
BenchmarkVectorizedBuiltinOpFunc/builtinUnaryMinusDecimalSig-VecBuiltinFunc-12         	  131503	      8770 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinOpFunc/builtinUnaryMinusDecimalSig-NonVecBuiltinFunc-12      	   27243	     43282 ns/op	   37872 B/op	     789 allocs/op
PASS
ok  	github.com/pingcap/tidb/expression	2.905s
```